# apps/backend/apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admins": {
      "get": {
        "description": "관리자 목록을 조회합니다. `status` 쿼리로 승인 상태로 필터링할 수 있습니다.",
        "operationId": "AdminsController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "승인 상태 (pending: 미승인, approved: 승인됨)",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "관리자 목록 조회",
        "tags": [
          "관리자"
        ]
      }
    },
    "/admins/{id}": {
      "get": {
        "description": "ID로 단일 관리자를 조회합니다.",
        "operationId": "AdminsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "관리자 상세 조회",
        "tags": [
          "관리자"
        ]
      },
      "patch": {
        "description": "관리자 정보를 수정합니다.",
        "operationId": "AdminsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "관리자 수정",
        "tags": [
          "관리자"
        ]
      },
      "delete": {
        "description": "관리자를 삭제합니다(소프트 삭제). MANAGER 이상 권한 필요. 본인은 삭제할 수 없습니다.",
        "operationId": "AdminsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "관리자 삭제",
        "tags": [
          "관리자"
        ]
      }
    },
    "/admins/{id}/approve": {
      "post": {
        "description": "승인 대기 중인 MEMBER 관리자를 승인합니다. MANAGER 이상 권한 필요.",
        "operationId": "AdminsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "관리자 승인",
        "tags": [
          "관리자"
        ]
      },
      "delete": {
        "description": "승인된 MEMBER 관리자를 다시 대기 상태로 되돌립니다. MANAGER 이상 권한 필요. 본인 승인은 회수할 수 없습니다.",
        "operationId": "AdminsController_revokeApproval",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "관리자 승인 회수",
        "tags": [
          "관리자"
        ]
      }
    },
    "/admins/{id}/role": {
      "patch": {
        "description": "대상 관리자의 권한을 MANAGER 또는 MEMBER로 변경합니다. OWNER 권한 필요. 본인은 변경할 수 없으며, MEMBER로 임명될 때 자동 승인됩니다.",
        "operationId": "AdminsController_assignRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "관리자 권한 임명",
        "tags": [
          "관리자"
        ]
      }
    },
...
    "/comments": {
      "post": {
        "description": "게시물에 댓글을 작성합니다. 상위 댓글 ID를 지정하면 대댓글로 등록됩니다. 비공개(PRIVATE) 공간 또는 게시판이라면 멤버여야 작성할 수 있습니다.",
        "operationId": "CommentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "댓글 작성",
        "tags": [
          "댓글"
        ]
      },
      "get": {
        "description": "게시물에 속한 댓글 목록을 조회합니다. `parentId`를 지정하지 않으면 해당 게시물의 최상위 댓글을, 지정하면 해당 댓글의 대댓글을 반환합니다. 비공개(PRIVATE) 공간 또는 게시판에 속한 댓글은 해당 공간/게시판의 멤버에게만 노출됩니다.",
        "operationId": "CommentsController_findAll",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "query",
            "description": "게시물 ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "parentId",
            "required": false,
            "in": "query",
            "description": "상위 댓글 ID. 지정하지 않으면 최상위 댓글을, 지정하면 해당 댓글의 대댓글을 반환합니다.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "댓글 목록 조회",
        "tags": [
          "댓글"
        ]
      }
    },
    "/comments/{id}": {
      "get": {
        "description": "ID로 단일 댓글을 조회합니다. 접근 권한이 없는 비공개(PRIVATE) 공간/게시판의 댓글은 조회되지 않습니다.",
        "operationId": "CommentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "댓글 상세 조회",
        "tags": [
          "댓글"
        ]
      },
      "patch": {
        "description": "댓글을 수정합니다. 본인이 작성한 댓글만 수정할 수 있습니다.",
        "operationId": "CommentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "댓글 수정",
        "tags": [
          "댓글"
        ]
      },
      "delete": {
        "description": "댓글을 삭제합니다(소프트 삭제). 작성자 본인이거나, 해당 게시판의 소유자(OWNER)/관리자(MANAGER)만 삭제할 수 있습니다.",
        "operationId": "CommentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "댓글 삭제",
        "tags": [
          "댓글"
        ]
      }
    }
  },
  "info": {
    "title": "blommunity API 문서",
    "description": "blommunity의 REST API 문서입니다. 인증, 게시물 관리 등의 기능을 제공합니다.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AdminEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "관리자 ID"
          },
          "name": {
            "type": "string",
            "description": "관리자 이름"
          },
          "email": {
            "type": "string",
            "description": "관리자 이메일"
          },
          "role": {
            "description": "관리자 권한",
            "enum": [
              "OWNER",
              "MANAGER",
              "MEMBER"
            ],
            "type": "string"
          },
          "approvedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "description": "승인 일시 (null이면 승인 대기)"
          },
          "approverId": {
            "type": "string",
            "nullable": true,
            "description": "승인자 ID"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "생성일시"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "수정일시"
          }
        },
        "description": "관리자 조회 응답에 포함되는 정보",
        "required": [
          "id",
          "name",
          "email",
          "role",
          "approvedAt",
          "approverId",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateAdminDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "관리자 이름"
          },
          "email": {
            "type": "string",
            "description": "이메일 주소",
            "format": "email"
          }
        },
        "description": "관리자 수정 요청 시 전달하는 정보 (모든 필드 선택)"
      },
      "AssignRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "description": "임명할 권한 (OWNER, MANAGER 또는 MEMBER)",
            "enum": [
              "OWNER",
              "MANAGER",
              "MEMBER"
            ]
          }
        },
        "description": "계정 권한 임명 요청 시 전달하는 정보",
        "required": [
          "role"
        ]
      },
...
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "postId": {
            "type": "string",
            "description": "소속 게시물 ID",
            "format": "uuid"
          },
          "parentId": {
            "type": "string",
            "nullable": true,
            "description": "상위 댓글 ID (대댓글인 경우)",
            "format": "uuid"
          },
          "content": {
            "type": "string",
            "description": "댓글 내용"
          }
        },
        "description": "댓글 생성 요청 시 전달하는 정보",
        "required": [
          "postId",
          "content"
        ]
      },
      "CommentEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "댓글 ID"
          },
          "postId": {
            "type": "string",
            "description": "소속 게시물 ID"
          },
          "parentId": {
            "type": "string",
            "nullable": true,
            "description": "상위 댓글 ID"
          },
          "authorId": {
            "type": "string",
            "description": "작성자 ID"
          },
          "content": {
            "type": "string",
            "description": "댓글 내용"
          },
          "replyCount": {
            "type": "number",
            "description": "활성 대댓글 수 (모든 depth 포함)"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "작성일시"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "수정일시"
          }
        },
        "description": "게시물에 속한 댓글의 상세 정보",
        "required": [
          "id",
          "postId",
          "parentId",
          "authorId",
          "content",
          "replyCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateCommentDto": {
        "type": "object",
        "properties": {
          "postId": {
            "type": "string",
            "description": "소속 게시물 ID",
            "format": "uuid"
          },
          "parentId": {
            "type": "string",
            "nullable": true,
            "description": "상위 댓글 ID (대댓글인 경우)",
            "format": "uuid"
          },
          "content": {
            "type": "string",
            "description": "댓글 내용"
          }
        },
        "description": "댓글 수정 요청 시 전달하는 정보 (모든 필드 선택)"
      }
    }
  }
}